# docs/openapi.yaml
openapi: 3.0.3
info:
  title: knowledge2exam API
  version: 1.0.0
  description: |
    试卷生成工具后端接口契约。REST + SSE，所有端点前缀 `/api/v1`。请求与响应均为 JSON，
    除上传接口用 `multipart/form-data`。

    本文件是前后端共用的机器可读契约，与后端 `app/schemas`、前端 `frontend/src/api/types.ts`
    保持同步。两个工作区（前端 / 后端）均可由此生成客户端 SDK 或服务端骨架。
servers:
  - url: /api/v1
    description: 相对路径（同源部署，经反向代理去掉前缀）
security: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: 刷新 access token 并轮换 refresh token（通过 Cookie）
      security: []
      responses:
        '200':
          description: 刷新成功，返回用户信息；新认证信息通过 httpOnly Cookie 传递
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: 当前用户任务列表
      description: 最近 50 条，按创建时间倒序。列表项不含 progress / plan.distribution 聚合
        （以详情端点为准）。
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsResponse'
    post:
      tags: [Jobs]
      operationId: createJob
      summary: 创建生成任务
      description: 立即返回 `job_id`（202），生成异步进行。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
      responses:
        '202':
          description: 任务已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      operationId: getJob
      summary: 任务快照（轮询降级用）
      description: 与 SSE 状态一致。`last_event_seq` 可用于 SSE 重连续传。
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 任务当前快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/JobNotFound'
    delete:
      tags: [Jobs]
      operationId: deleteJob
      summary: 删除任务及产物
      description: 删除任务及其阶段记录与存储产物；已共享资源不随之删除。
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: 已删除
        '404':
          $ref: '#/components/responses/JobNotFound'
  /jobs/{job_id}/events:
    get:
      tags: [Jobs]
      operationId: streamJobEvents
      summary: SSE 进度流
      description: |
        `Content-Type: text/event-stream`。每帧含 `id`（对应 `job_stage.seq`）、`event`、`data`。
        客户端断连后用 `Last-Event-ID` 头重连，服务端从该 `seq` 之后补发。

        事件类型与 `data` 结构（由 `app/orchestration/stages.py` 发出）：

        - `stage_changed`：`{stage, previous}` —— 阶段切换（preprocessing / generating / rendering）；
          首帧 `previous` 为 `pending`，后续帧为上一个 Stage
        - `plan_ready`：`{total, distribution, reference_used, duration_minutes}` ——
          agent 出题蓝图就绪；`distribution` 按题型（choice / blank / short_answer）计数，
          `reference_used` 当前恒为 `agent_planning`
        - `progress`：`{completed, total}` —— 蓝图驱动的出题进度（todo 状态统计）
        - `warning`：`{code, upload_id?, message}` —— 非致命告警；
          `code` ∈ PARSE_FAILED / OCR_DEGRADED / AGENT_WARNING / RENDER_FAILED
        - `error`：`{error_code, message}` —— 管线意外异常，任务定格 failed
        - `revision_started`：`{round_no, feedback}` —— 一轮划选反馈修订开始
        - `revision_done`：`{round_no, render_status}` —— 修订完成，试卷已更新并重渲染
        - `revision_failed`：`{round_no, message}` —— 修订失败（agent 异常或未产生修改）
        - `done`：`{status, total?, abandoned?, md_url?, pdf_url?}` —— 终态关闭事件；
          `status` ∈ completed / partially_completed / failed / cancelled（取消时仅含 status）
      parameters:
        - $ref: '#/components/parameters/JobId'
        - name: Last-Event-ID
          in: header
          required: false
          description: 断线重连续传：服务端从该 seq 之后补发事件
          schema:
            type: integer
      responses:
        '200':
          description: SSE 事件流
          content:
            text/event-stream:
              schema:
                type: string
                description: "按 `id: <seq>\\nevent: <type>\\ndata: <json>\\n\\n` 帧序列"
        '404':
          $ref: '#/components/responses/JobNotFound'
  /jobs/{job_id}/paper.md:
    get:
      tags: [Jobs]
      operationId: downloadPaperMarkdown
      summary: 获取 md 试卷（含进行中预览）
      description: |
        agent 整卷直写 output/paper.md；任务进行中也可调用，返回当前正在书写的稿子
        （md_key 尚未提交时回退读取工作区文件），支撑前端渐进预览。
        题卷在前，答案与解析单独成篇。
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Markdown 试卷
          content:
            text/markdown:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/JobNotFound'
        '409':
          $ref: '#/components/responses/JobNotReady'
  /jobs/{job_id}/paper.pdf:
    get:
      tags: [Jobs]
      operationId: downloadPaperPdf
      summary: 下载 PDF 试卷
      description: |
        任务未到 `rendering` 完成时返回 409；`partially_completed` 且 PDF 渲染失败时返回 409 `RENDER_FAILED`。
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: PDF 试卷
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/JobNotFound'
        '409':
          $ref: '#/components/responses/JobNotReady'
  /jobs/{job_id}/cancel:
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: 取消运行中的任务
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: 已发出取消
        '404':
          $ref: '#/components/responses/JobNotFound'
        '409':
          $ref: '#/components/responses/JobAlreadyFinished'
  /jobs/{job_id}/revisions:
    post:
      tags: [Revisions]
      operationId: createRevision
      summary: 提交一轮划选反馈修订
      description: |
        用户在试卷预览中划选一部分内容并提出要求；请求交回完成该任务的 agent
        （同一全套工具，携带划选锚点、反馈与全部历史轮次）直接续跑修改
        output/paper.md 并重渲染。试卷内容直接生效，修订前后快照留存于
        paper_revision 表。仅限 completed / partially_completed 任务，且同任务
        同时只能有一轮修订在跑。
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionCreate'
      responses:
        '202':
          description: 修订已受理，后台续跑
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionAccepted'
        '400':
          description: 载荷不合法（如 feedback 为空）
        '404':
          $ref: '#/components/responses/JobNotFound'
        '409':
          description: 任务未完成（JOB_NOT_FINISHED）或已有修订在跑（REVISION_IN_PROGRESS）
    get:
      tags: [Revisions]
      operationId: listRevisions
      summary: 修订会话历史
      description: 该任务的全部修订轮次（新→旧），即修订会话记录。
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 修订历史
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionsResponse'
        '404':
          $ref: '#/components/responses/JobNotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        Cookie 认证。登录成功后，后端设置 httpOnly Cookie `access_token`（JWT，15 分钟）与
        `refresh_token`（不透明令牌，30 天）。前端请求需携带 `withCredentials: true`，
        浏览器会自动附带 Cookie。刷新与登出同样通过 Cookie 传递 refresh_token。
  parameters:
...
    JobId:
      name: job_id
      in: path
      required: true
      description: 任务 ID
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 请求参数或载荷不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    JobNotFound:
      description: 任务不存在或无权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    JobNotReady:
      description: 产物尚未生成，或 PDF 渲染失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    JobAlreadyFinished:
      description: 任务已终结，不可取消
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: 未携带有效 access token，或 refresh token 已吊销/过期/重放
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    User:
      type: object
      required: [user_id, email, username]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    JobCreate:
      type: object
      description: 创建任务请求
      properties:
        upload_ids:
          type: array
          items:
            type: string
            format: uuid
          default: []
          description: 本次使用的上传件；为空且无任何文本类上传件时返回 400 INPUT_EMPTY
        school_id:
          type: string
          format: uuid
          nullable: true
          description: 有共享项或需读共享库时必填
        course_id:
          type: string
          format: uuid
          nullable: true
          description: 同上
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 300
          default: 100
          description: 考试时长
        need_explanation:
          type: boolean
          default: false
          description: 是否生成解析
    JobAccepted:
      type: object
      required: [job_id, status, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
        created_at:
          type: string
          format: date-time
          description: ISO 8601 UTC，带 Z
    Plan:
      type: object
      description: 规划概要
      properties:
        total:
          type: integer
        distribution:
          type: object
          properties:
            choice:
              type: integer
            blank:
              type: integer
            short_answer:
              type: integer
    Progress:
      type: object
      properties:
        completed:
          type: integer
        total:
          type: integer
        retried:
          type: integer
        replanned:
          type: integer
        abandoned:
          type: integer
    Warning:
      type: object
      properties:
        code:
          type: string
          enum: [PARSE_FAILED, OCR_DEGRADED, AGENT_WARNING, RENDER_FAILED]
        upload_id:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
    Artifacts:
      type: object
      description: 产物下载地址
      properties:
        md_url:
          type: string
          nullable: true
          description: 相对路径，如 /api/v1/jobs/{job_id}/paper.md
        pdf_url:
          type: string
          nullable: true
    Job:
      type: object
      description: 任务快照
      required: [job_id, status, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        stage:
          $ref: '#/components/schemas/Stage'
        duration_minutes:
          type: integer
        need_explanation:
          type: boolean
        plan:
          $ref: '#/components/schemas/Plan'
        progress:
          $ref: '#/components/schemas/Progress'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/Warning'
        artifacts:
          $ref: '#/components/schemas/Artifacts'
        last_event_seq:
          type: integer
          description: SSE 续传游标
        error_code:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
    JobsResponse:
      type: object
      description: 当前用户任务列表（GET /jobs）
      required: [jobs]
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
    JobStatus:
      type: string
      enum: [pending, preprocessing, generating, rendering, completed, partially_completed, failed, cancelled]
      description: |
        任务状态机（见 app/orchestration/state_machine.py）：pending → preprocessing →
        generating → rendering → completed；rendering 亦可转入 partially_completed（PDF 渲染
        失败，仅交付 Markdown）；任一运行态可转入 failed / cancelled。
    Stage:
      type: string
      enum: [preprocessing, generating, rendering]
    SelectionAnchor:
      type: object
      description: 用户在试卷预览中划选的锚点：选中文本 + 前后文（内容定位，非偏移）。
      required: [text]
      properties:
        text:
          type: string
          description: 用户划选的文本
        before:
          type: string
          description: 选区前文（少量上下文）
        after:
          type: string
          description: 选区后文（少量上下文）
    RevisionCreate:
      type: object
      description: 提交一轮修订：划选位置与内容 + 用户要求。
      required: [selection, feedback]
      properties:
        selection:
          $ref: '#/components/schemas/SelectionAnchor'
        feedback:
          type: string
          description: 用户对所选部分的修订要求
    RevisionAccepted:
      type: object
      required: [revision_id, round_no, status]
      properties:
        revision_id:
          type: string
          format: uuid
        round_no:
          type: integer
          description: 轮次号，从 1 递增
        status:
          type: string
          enum: [running]
    RevisionItem:
      type: object
      description: 修订历史条目（会话记录）。快照不随列表返回。
      required: [revision_id, round_no, status, feedback, created_at]
      properties:
        revision_id:
          type: string
          format: uuid
        round_no:
          type: integer
        status:
          type: string
          enum: [running, done, failed]
        selection:
          allOf:
            - $ref: '#/components/schemas/SelectionAnchor'
          nullable: true
        feedback:
          type: string
        summary:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        applied_at:
          type: string
          format: date-time
          nullable: true
    RevisionsResponse:
      type: object
      required: [job_id, total, revisions]
      properties:
        job_id:
          type: string
          format: uuid
        total:
          type: integer
        revisions:
          type: array
          items:
            $ref: '#/components/schemas/RevisionItem'
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
          description: 结构化的补充信息，随错误码不同而异
    ErrorCode:
      type: string
      description: |
        错误码全集（与 app/core/exceptions.py 的 ErrorCode 一致）。含仅作为 warning / error
        事件出现、无 HTTP 状态的码（如 PARSE_FAILED、AGENT_WARNING、PIPELINE_FAILED）。
      enum:
        - INPUT_EMPTY
        - UNSUPPORTED_FORMAT
        - FILE_TOO_LARGE
        - PDF_ENCRYPTED
        - SHARE_NOT_ALLOWED
        - SHARE_SCOPE_REQUIRED
        - INVALID_DURATION
        - JOB_NOT_FOUND
        - SCHOOL_NOT_FOUND
        - UPLOAD_IN_USE
        - JOB_NOT_READY
        - JOB_ALREADY_FINISHED
        - JOB_NOT_FINISHED
        - REVISION_IN_PROGRESS
        - PAPER_NOT_READY
        - PARSE_FAILED
        - OCR_DEGRADED
        - MODERATION_REJECTED
        - PLANNING_FAILED
        - GENERATION_EXHAUSTED
        - RENDER_FAILED
        - AGENT_WARNING
        - PIPELINE_FAILED
        - SERVER_RESTARTED
        - MODEL_UNAVAILABLE
        - RATE_LIMITED
        - USER_EXISTS
        - INVALID_CREDENTIALS
        - TOKEN_MISSING
        - TOKEN_EXPIRED
        - REFRESH_INVALID
        - REFRESH_REUSE_DETECTED